rename socialsmodal file
