+++ TrashMob/client-app/src/components/EventManagement/ShareToSocialsModal.tsx
